Update access-control form wrappers and tests.
Add a permissions-wrapped version of 'grouped_collection_select'.

Also, a little mummery to keep other tests passing.  There's a
new junk method in the form helpers class which we don't have to wrap.
Also, they'd switched the order in which the checkbox itself and the
associated hidden input show up; had to adjust test assertions
to compensate.

git-svn-id: svn+ssh://svn.int.smartleaf.com/home/svn/smartleaf/trunk@5009 44e4a855-3122-4fc0-a98e-401d94ebb483