Avoid redundant permission checks on associations.
When an association proxy gets set, the underlying foreign key attribute
gets set as well.  Smartguard was doing access control checks for both
--- the second of which was really slow, as the foreign key check no
longer had access to the AR object containing the access control keys,
and so had to go to the database.

So, the Smartguard belongs_to wrapper now sets an attribute to mark
that "we've already checked this object" while executing the underlying
association code (using begin ... ensure to clean up afterward), and
the code that handles simple attributes looks aside to see if this
is a redundant foreign key check, and if so, doesn't repeat the work.

git-svn-id: svn+ssh://svn.int.smartleaf.com/home/svn/smartleaf/trunk@5185 44e4a855-3122-4fc0-a98e-401d94ebb483