Work around limitations on Oracle support of recursive "with ...".
In particular, a select with a "with ..." clause prefix can't appear
as a branch of a union, so we move the union inside.  There are two
other documented limitations covered by the error I was getting:
limitations on nesting, though I seem to be getting away with it here,
and also "cannot appear in parentheses".  That last one is astonishing,
but verifiable:  take a working "with ... select", surround it in
parens, type it in to raw SQL plus, and you get an ORA-32034.