Have Gemfile check if Oracle is set up; *don't* save Gemfile.lock.
We want tests to be runnable even if the Oracle client code is not
present.  But the Oracle client code is needed to build the binary
extensions in the Oracle extended adapter --- so, on a Postgres-only
system, that adapter gem can't even be in the Gemfile.

The only straightforward way to arrange that is to put
conditionalization in the Gemfile.  But if we're doing that, we
can't freeze the bundle configuration in Gemfile.lock.  So, it gets
generated at "bundle install" every time.  There are obvious problems
with that, but not being able to run the tests at all on systems
without Oracle would be a great deal worse.

(Arrangement recommended by Prem Sichanugrist at a bostonrb project
night; I don't think either of us is really thrilled with it, but
I can't think of anything better.)