Allow attr_write_access_checks to be bypassed when copying a record.
Our deep_copy code wants to set all attributes on a new record to
the values of some pre-existing one -- even the set-guarded ones.
Arguably, it should be cloning the record, but it used to tweak Rails
internals to effectively bypass the assignment path.  This creates a
targeted method which allows attr access checks to be bypassed for an
*unsaved* record only...