Make our 'assign_attributes' wrapper enforce strong parameters properly.
Before, it would eventually raise the ForbiddenAttributesException, but
might set a few of them first (if they were listed in block-set groups).
This seems generally undesirable.  Regrettably, this works by invoking
undocumented internals, but this particular thing seems unlikely to
change, and it is a one-liner.

Tossed in a test for the block-set-group functionality as well, as I
couldn't find an earlier one.