Use simpler (but more varied) SQL for 'ids_permitting'.
In this variant, rather than joining to the permissions table
(in a way that's likely to be *very* non-transparent to query
optimizers), we generate one

  (ack_col = ack_value and ack2_col = ack2_value and ...)

for each permission the user actually has, and then OR them
together if there are more than one.  This eliminates the
recursive queries that make it hard for the optimizer to
guess how many perms a given user has (and easy to wildly
overestimate), and gives it a fighting chance at guessing
how many permitted rows there actually are.